feat: modify user_based model corresponding refactor pipeline
