+++ recommender/preprocess/preprocess_movielens.py
@@ -2,7 +2,7 @@ from typing import Dict, Union
 
 import pandas as pd
 
-from recommender.preprocess.preprocess_base_ import PreoprocessorBase
+from recommender.preprocess.preprocess_base import PreoprocessorBase
 
 
 class Preprocessor(PreoprocessorBase):
